modifier ,supprimer catagorie

+++ src/app/services/categorie/categorie.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import {Observable, of} from 'rxjs';
 import { CategorieDto } from 'src/app/dto/categorie-dto';
 import { Categorie } from 'src/app/models/categirie';
 
@@ -13,33 +13,53 @@ export class CategorieService {
 
   constructor(private http: HttpClient) { }
 
-  // Méthode pour récupérer un article
+  /**
+   *  service pour récupérer une categorie
+   * @param id
+   */
   getCategorie(id: number): Observable<Object> {
     const url =  this.baseUrl+`/${id}`;
     return this.http.get(url);
   }
 
-  // Méthode pour ajouter un article
+  /**
+   * service pour ajouter une categorie
+   * @param categirie
+   */
   ajouterCategorie(categirie: Categorie): Observable<CategorieDto> {
     const url =  this.baseUrl;
     return this.http.post(url, categirie);
   }
 
-  // Méthode pour mettre à jour un article
+  /**
+   *   Service pour mettre à jour une categorie
+   * @param id
+   * @param categirie
+   */
   mettreAJourCategorie(id: number, categirie: any): Observable<any> {
     const url =  this.baseUrl+`/${id}`;
     return this.http.put(url, categirie);
   }
 
-  // Méthode pour supprimer un article
-  supprimerCategorie(id: number): Observable<Object> {
-    const url =  this.baseUrl+`/${id}`;
-    return this.http.delete(url);
+  /**
+   *  Service pour supprimer une categorie
+   * @param id
+   */
+  supprimerCategorie(id?: number): Observable<Object> {
+    if(id){
+      const url =  this.baseUrl+`/${id}`;
+      return this.http.delete(url);
+    }
+   return of();
   }
 
-  // Méthode pour récupérer toutes les articles
+  /**
+   *    Service pour récupérer toutes les categories
+   */
   getToutesCategories(): Observable<any> {
     const url =  this.baseUrl;
     return this.http.get(url);
   }
+
+
 }
